move Metamaps code into webpack
