+++ app/assets/javascripts/application.js
@@ -13,37 +13,11 @@
 //= require jquery
 //= require jquery-ui
 //= require jquery_ujs
-//= require ./webpacked/metamaps.bundle
 //= require_directory ./lib
-//= require ./src/Metamaps.GlobalUI
-//= require ./src/Metamaps.Router
-//= require ./src/Metamaps.Backbone
-//= require ./src/Metamaps.Views
+//= require ./src/JIT
+//= require ./src/Metamaps.Erb
+//= require ./webpacked/metamaps.bundle
 //= require ./src/views/chatView
 //= require ./src/views/videoView
 //= require ./src/views/room
-//= require ./src/JIT
 //= require ./src/check-canvas-support
-//= require ./src/Metamaps
-//= require ./src/Metamaps.Create
-//= require ./src/Metamaps.TopicCard
-//= require ./src/Metamaps.SynapseCard
-//= require ./src/Metamaps.Visualize
-//= require ./src/Metamaps.Util
-//= require ./src/Metamaps.Realtime
-//= require ./src/Metamaps.Control
-//= require ./src/Metamaps.Filter
-//= require ./src/Metamaps.Listeners
-//= require ./src/Metamaps.Organize
-//= require ./src/Metamaps.Topic
-//= require ./src/Metamaps.Synapse
-//= require ./src/Metamaps.Map
-//= require ./src/Metamaps.Account
-//= require ./src/Metamaps.Mapper
-//= require ./src/Metamaps.Mobile
-//= require ./src/Metamaps.Admin
-//= require ./src/Metamaps.Import
-//= require ./src/Metamaps.AutoLayout
-//= require ./src/Metamaps.PasteInput
-//= require ./src/Metamaps.JIT
-//= require ./src/Metamaps.Debug
