+++ app/assets/javascripts/src/Metamaps.Erb.js.erb
@@ -1,0 +1,18 @@
+/* global Metamaps */
+
+/*
+ * Metamaps.Erb.js.erb
+ */
+
+/* erb variables from rails */
+window.Metamaps = window.Metamaps || {}
+Metamaps.Erb = {}
+Metamaps.Erb['REALTIME_SERVER'] = '<%= ENV['REALTIME_SERVER'] %>'
+Metamaps.Erb['junto_spinner_darkgrey.gif'] = '<%= asset_path('junto_spinner_darkgrey.gif') %>'
+Metamaps.Erb['user.png'] = '<%= asset_path('user.png') %>'
+Metamaps.Erb['icons/wildcard.png'] = '<%= asset_path('icons/wildcard.png') %>'
+Metamaps.Erb['topic_description_signifier.png'] = '<%= asset_path('topic_description_signifier.png') %>'
+Metamaps.Erb['topic_link_signifier.png'] = '<%= asset_path('topic_link_signifier.png') %>'
+Metamaps.Erb['synapse16.png'] = '<%= asset_path('synapse16.png') %>'
+Metamaps.Metacodes = <%= Metacode.all.to_json.gsub(%r[(icon.*?)(\"},)], '\1?purple=stupid\2').html_safe %>
+Metamaps.VERSION = '<%= METAMAPS_VERSION %>'
