+++ frontend/src/Metamaps/Metamaps.AutoLayout.js
@@ -1,3 +1,4 @@
+window.Metamaps = window.Metamaps || {}
 /* global Metamaps */
 
 /*
