+++ frontend/src/Metamaps/Metamaps.Backbone.js
@@ -1,3 +1,4 @@
+window.Metamaps = window.Metamaps || {}
 /* global Metamaps, Backbone, _, $ */
 
 /*
