+++ frontend/src/Metamaps/Metamaps.Control.js
@@ -1,3 +1,4 @@
+window.Metamaps = window.Metamaps || {}
 /* global Metamaps, $ */
 
 /*
