+++ frontend/src/Metamaps/Metamaps.PasteInput.js
@@ -1,3 +1,4 @@
+window.Metamaps = window.Metamaps || {}
 /* global Metamaps, $ */
 
 /*
