+++ frontend/src/Metamaps/Metamaps.ReactComponents.js
@@ -1,0 +1,7 @@
+window.Metamaps = window.Metamaps || {}
+
+import Maps from '../components/Maps'
+
+Metamaps.ReactComponents = {
+  Maps
+}
