+++ frontend/src/Metamaps/Metamaps.Realtime.js
@@ -1,3 +1,5 @@
+window.Metamaps = window.Metamaps || {}
+
 /* global Metamaps, $ */
 
 /*
