+++ frontend/src/Metamaps/Metamaps.Visualize.js
@@ -1,3 +1,4 @@
+window.Metamaps = window.Metamaps || {}
 /* global Metamaps, $ */
 /*
  * Metamaps.Visualize
